Add some protocols and extensions
