--- GoogleMapsTestApp/Reusables/Struct/ButtonSettings.swift
@@ -1,14 +1,0 @@
-//
-//  ButtonSettings.swift
-//  GoogleMapsTestApp
-//
-//  Created by Andrey on 15.11.2019.
-//  Copyright © 2019 Andrey. All rights reserved.
-//
-
-import Foundation
-
-struct ButtonSettings {
-    static let cornerRadius: CGFloat = 20
-    static let color = UIColor.white
-}
